% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{as.character.card_vector}
\alias{as.character.card_vector}
\title{Coercion Method for \code{card_vector} to \code{character}}
\usage{
\method{as.character}{card_vector}(x, ...)
}
\arguments{
\item{x}{A \code{card_vector} object.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A character vector.
}
\description{
This function converts a \code{card_vector} to a character vector by extracting
the underlying character data.
}
\examples{
# Coerce card_vector to character
deck <- card_vector(c("A♠", "10♣"))
character_deck <- as.character(deck)
character_deck
}
