% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{card_is_face}
\alias{card_is_face}
\title{Check if a card is a face card (J, Q, K)}
\usage{
card_is_face(x)
}
\arguments{
\item{x}{A card_vector}
}
\value{
Logical vector
}
\description{
Check if a card is a face card (J, Q, K)
}
