% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{card_rank}
\alias{card_rank}
\title{Get card ranks}
\usage{
card_rank(x)
}
\arguments{
\item{x}{A card_vector}
}
\value{
Character vector of ranks (e.g., "A", "10", "Q")
}
\description{
Get card ranks
}
