% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{card_suit}
\alias{card_suit}
\title{Get Card Suits}
\usage{
card_suit(x)
}
\arguments{
\item{x}{A \code{card_vector} object.}
}
\value{
A character vector containing the suits of the cards (e.g., "♠", "♥").
}
\description{
Extracts the suit (♠, ♥, ♦, ♣) of each card in the \code{card_vector}.
}
\examples{
# Extract card suits from a deck
deck <- card_vector(c("A♠", "10♣", "K♥"))
card_suit(deck)
}
