% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{card_vector}
\alias{card_vector}
\title{Card Vector Constructor}
\usage{
card_vector(x = character())
}
\arguments{
\item{x}{A character vector of cards like "A♠", "10♣", "K♥"}
}
\value{
A card_vector object
}
\description{
Create a custom vctrs card vector.
}
