% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_board.R
\name{create_board}
\alias{create_board}
\title{Create a Shuffled Deck of Cards}
\usage{
create_board(num_decks = 4)
}
\arguments{
\item{num_decks}{The number of decks shuffled in the game}
}
\value{
A shuffled deck of cards.
}
\description{
Create a Shuffled Deck of Cards
}
\examples{
# Create a single shuffled deck
deck1 <- create_board(1)
head(deck1)

# Create and view number of cards in 6 decks
deck6 <- create_board(6)
length(deck6)  # Should be 312

}
