% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_cards.R
\name{deal_cards}
\alias{deal_cards}
\title{Deal Cards to Players and Dealer}
\usage{
deal_cards(num_players, deck, split = FALSE)
}
\arguments{
\item{num_players}{Integer. Number of players in the game.}

\item{deck}{A vector representing the shuffled deck of cards.}

\item{split}{Logical. If TRUE, the function allows for splitting hands.}
}
\value{
A list containing the player hands, dealer hand, and remaining deck.
}
\description{
This function deals cards to the specified number of players and a dealer.
Each player receives two cards, and the dealer receives two cards.
Optionally, hands can be split.
}
