% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealer_play.R
\name{dealer_play}
\alias{dealer_play}
\title{Dealer's play logic according to Blackjack rules}
\usage{
dealer_play(dealer_hand, deck)
}
\arguments{
\item{dealer_hand}{Character vector of the dealer's current hand}

\item{deck}{Character vector of the remaining deck}
}
\value{
A list with updated dealer_hand and deck
}
\description{
Dealer draws cards until reaching at least 17, or a 5-card non-bust.
}
