% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_bet.R
\name{insurance_bet}
\alias{insurance_bet}
\title{Offer and Resolve Insurance Bet}
\usage{
insurance_bet(dealer_hand, player_accepts)
}
\arguments{
\item{dealer_hand}{A character vector of two cards (e.g., c("A♠", "10♦"))}

\item{player_accepts}{Logical, TRUE if player accepts insurance}
}
\value{
A list with insurance_available, insurance_paid, payout, blackjack
}
\description{
If the dealer's upcard is an Ace, allows the player to take insurance.
Insurance pays 2:1 if the dealer has a Blackjack.
}
\examples{
insurance_bet(c("A♠", "10♣"), player_accepts = TRUE)
insurance_bet(c("A♠", "8♣"), player_accepts = TRUE)
insurance_bet(c("9♠", "10♣"), player_accepts = TRUE)

}
