% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_result.R
\name{new_round_result}
\alias{new_round_result}
\title{Create a round_result object}
\usage{
new_round_result(player, hand, score, outcome)
}
\arguments{
\item{player}{Character: player name}

\item{hand}{Character vector: player's hand}

\item{score}{Integer: final score}

\item{outcome}{Character: outcome ("Win", "Lose", "Tie", "Surrender")}
}
\value{
An object of class \code{round_result}
}
\description{
Create a round_result object
}
