% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_blackjack.R
\name{play_blackjack}
\alias{play_blackjack}
\title{Play a Full Game of Blackjack Interactively}
\usage{
play_blackjack(rule = "standard")
}
\description{
Asks user for number of players and decks, and plays a full round of Blackjack
}
