% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hand.R
\name{score_hand}
\alias{score_hand}
\title{Score the Cards in a Hand}
\usage{
score_hand(hand)
}
\arguments{
\item{hand}{A character vector of cards.}
}
\value{
An integer representing the total score.
}
\description{
Calculates the total score for a Blackjack hand. Aces count as 11 unless it causes a bust.
}
\examples{
score_hand(c("A♠", "10♣"))  # 21
score_hand(c("A♠", "A♦", "9♠"))  # 21
score_hand(c("A♠", "A♦", "A♣", "10♠"))  # 13
}
