% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{vec_cast.card_vector.character}
\alias{vec_cast.card_vector.character}
\title{Coercion Method for \code{card_vector} to \code{character} for later use}
\usage{
\method{vec_cast}{card_vector.character}(x, to, ...)
}
\arguments{
\item{x}{A \code{card_vector} object.}

\item{to}{The target class to cast to (character).}

\item{...}{Additional arguments for method dispatch.}
}
\value{
A character vector.
}
\description{
Coercion method to cast a \code{card_vector} to a \code{character} vector.
}
\examples{
# Coerce card_vector to character
deck <- card_vector(c("A♠", "10♣"))
character_deck <- vctrs::vec_cast(deck, to = "character")
character_deck
}
