% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{vec_cast.character.card_vector}
\alias{vec_cast.character.card_vector}
\title{Coercion Method for \code{character} to \code{card_vector}}
\usage{
\method{vec_cast}{character.card_vector}(x, to, ...)
}
\arguments{
\item{x}{A \code{character} vector.}

\item{to}{The target class to cast to (\code{card_vector}).}

\item{...}{Additional arguments for method dispatch.}
}
\value{
A \code{card_vector} object.
}
\description{
Converts a \code{character} vector to a \code{card_vector}.
}
\examples{
# Coerce character vector to card_vector
character_deck <- c("A♠", "10♣")
deck <- vctrs::vec_cast(character_deck, to = "card_vector")
deck
}
