% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_winner.R
\name{announce_winner}
\alias{announce_winner}
\title{Announce Winner Function}
\usage{
announce_winner(player_scores, dealer_score)
}
\arguments{
\item{player_scores}{Named numeric vector of player scores.}

\item{dealer_score}{Numeric dealer score.}
}
\value{
Named character vector of outcomes ("Win", "Lose", "Tie").
}
\description{
This function announces the winner(s) of the game based on hand scores.
}
