% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{card_rank}
\alias{card_rank}
\title{Get Card Ranks}
\usage{
card_rank(x)
}
\arguments{
\item{x}{A \code{card_vector} object.}
}
\value{
A character vector containing the ranks of the cards.
}
\description{
Extracts the rank (number or face value) of each card in the \code{card_vector}.
}
\examples{
# Extract card ranks from a deck
deck <- card_vector(c("A♠", "10♣", "K♥"))
card_rank(deck)
}
