% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_board.R
\name{create_board}
\alias{create_board}
\title{Create a Shuffled Deck of Cards}
\usage{
create_board(num_decks = 4)
}
\arguments{
\item{num_decks}{The number of decks shuffled in the game}
}
\value{
A shuffled deck of cards.
}
\description{
Create a Shuffled Deck of Cards
}
\examples{

create_board()
}
