% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_cards.R
\name{deal_cards}
\alias{deal_cards}
\title{Deal Cards to Players and Dealer}
\usage{
deal_cards(num_players, deck)
}
\arguments{
\item{num_players}{Integer. The number of players in the game.}

\item{deck}{A vector representing the shuffled deck.}
}
\value{
A list with hands for players and dealer.
}
\description{
Deal Cards to Players and Dealer
}
\examples{
deck <- create_deck()
game <- deal_cards(2, deck)
}
