% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deal_cards.R
\name{deal_cards}
\alias{deal_cards}
\title{Deal Cards to Players and Dealer}
\usage{
deal_cards(num_players, deck)
}
\arguments{
\item{num_players}{Number of players}

\item{deck}{A shuffled deck of cards}
}
\value{
A list with player_hands, dealer_hand (1 card), and remaining deck
}
\description{
Deals 2 cards to each player and 1 card to the dealer. Dealer's 2nd card comes later.
}
