% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_down.R
\name{double_down}
\alias{double_down}
\title{Perform a double down action for a player}
\usage{
double_down(hand, deck)
}
\arguments{
\item{hand}{Character vector of player hand}

\item{deck}{Character vector of remaining deck}
}
\value{
A list with updated hand, updated deck, and validity flag
}
\description{
Player adds exactly one card to their initial 2-card hand.
}
