% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hand.R
\name{score_hand}
\alias{score_hand}
\title{Score the Cards in a Hand}
\usage{
score_hand(hand)
}
\arguments{
\item{hand}{A vector of cards.}
}
\value{
The score of the hand.
}
\description{
Score the Cards in a Hand
}
\examples{
# Score a hand with Ace and 9
score_hand(c("A", "9"))

# Score a hand with multiple Aces
score_hand(c("A", "A", "9"))

# Score a hand with multiple Aces and a 10, resulting in a bust
score_hand(c("A", "A", "A", "10"))
}
