% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hand_dynamic.R
\name{score_hand_dynamic}
\alias{score_hand_dynamic}
\title{Score the Cards in a Hand with Dynamic Rules}
\usage{
score_hand_dynamic(hand, rule = "standard")
}
\arguments{
\item{hand}{A vector of cards.}

\item{rule}{A character string indicating the scoring rule to use.
Options: "standard" (default), "custom". See details below.}
}
\value{
The score of the hand.
}
\description{
This function calculates the score of a hand based on dynamic rules.
By default, it uses the standard Blackjack rules, but it can be customized to
use other rules, such as always treating Aces as 1, or applying custom scoring mechanisms.
}
\examples{
# Score a hand with Ace and 9 (standard rule)
score_hand_dynamic(c("A♠", "9♣"))

# Score a hand with multiple Aces using the custom rule (Ace is always 1)
score_hand_dynamic(c("A♠", "A♦", "9♠"), rule = "custom")

# Score a hand with multiple Aces and a 10, resulting in a bust (standard rule)
score_hand_dynamic(c("A♠", "A♦", "A♣", "10♠"))

}
