% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_vector.R
\name{vec_cast.card_vector.card_vector}
\alias{vec_cast.card_vector.card_vector}
\title{Coercion Method for \code{card_vector}}
\usage{
\method{vec_cast}{card_vector.card_vector}(x, to, ...)
}
\arguments{
\item{x}{A \code{card_vector} object.}

\item{to}{The target class to cast to.}

\item{...}{Additional arguments for method dispatch.}
}
\value{
The original \code{card_vector} object, as no conversion is needed.
}
\description{
Coercion method for casting one \code{card_vector} to another.
}
\examples{
# Coerce between two card_vectors
deck1 <- card_vector(c("A♠", "10♣"))
deck2 <- vec_cast(deck1, to = "card_vector")
deck2
}
